server/requirements.txt
aiosmtpd==1.4.2
altgraph==0.17.2
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
atomicwrites==1.4.1
atpublic==3.0.1
attrs==22.1.0
auto-py-to-exe==2.18.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.4
coverage==5.5
cryptography==37.0.4
cssselect==1.1.0
cssutils==2.5.1
cycler==0.11.0
decorator==4.0.11
distlib==0.3.2
Django==4.1rc1
django-cors-headers==3.11.0
django-jet-4==0.0.1
django-utils-six==2.0
djangorestframework
djangorestframework-expiring-authtoken==0.1.4
djangorestframework-simplejwt
djoser==2.1.0
Eel==0.12.4
et-xmlfile==1.1.0
executing==0.8.3
filelock==3.0.12
Flask==2.1.3
fonttools==4.31.2
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
googletrans==3.1.0a0
greenlet==1.1.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
huepy==1.2.1
hyperframe==5.2.0
idna==2.10
imageio==2.21.1
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
iniconfig==1.1.1
instaloader==4.9.2
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
kiwisolver==1.4.2
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mock==4.0.3
MouseInfo==0.1.3
moviepy==1.0.3
mysqlclient==2.0.3
numpy==1.21.1
openpyxl==3.0.9
outcome==1.2.0
packaging==21.3
pandas==1.4.1
parso==0.8.3
pbr==5.6.0
pefile==2021.9.3
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2022.1.8
platformdirs==2.2.0
pluggy==1.0.0
premailer==3.10.0
proglog==0.1.10
prompt-toolkit==3.0.28
pure-eval==0.2.2
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.11.2
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.1
pywhatkit==5.4
pywin32-ctypes==0.2.0
requests==2.28.1
requests-toolbelt==0.9.1
responses==0.21.0
rfc3986==1.5.0
rsa==4.9
schedule==1.1.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
stack-data==0.2.0
stevedore==3.3.0
telebot==0.0.4
Telethon==1.24.0
termcolor==1.1.0
tomli==2.0.1
tqdm==4.64.0
traitlets==5.1.1
trio==0.21.0
trio-websocket==0.9.2
ttkSimpleDialog==1.2.1
tzdata==2022.1
UliPlot==0.2.2
unicode==2.8
Unidecode==1.3.4
urllib3==1.26.11
views==0.3
virtualenv==20.7.2
virtualenv-clone==0.5.6
virtualenvwrapper==4.8.4
wcwidth==0.2.5
Werkzeug==2.2.1
whichcraft==0.6.1
wikipedia==1.4.0
workbook==1.1
wsproto==1.1.0
xlrd==2.0.1
XlsxWriter==3.0.3
xlutils==2.0.0
xlwt==1.3.0
yagmail==0.15.280
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
